backend/requirements.txt
Flask==2.3.3
Flask-CORS==4.0.0
SQLAlchemy==2.0.21
python-dotenv==1.0.0
psycopg2-binary==2.9.7
alembic==1.12.0